+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 1.53
 ====
 * Added the count option to the bare interface (Iss #45)
+* Added the valuereport equivalent to REST and bare interface (Iss #44)
 
 1.52
 ====